Working constraints converter along with a much more robust strategy-explorer.
