+++ runtime/test/planner-tests.js
@@ -13,8 +13,10 @@ let Arc = require('../arc.js');
 let Loader = require('../loader.js');
 let Planner = require('../planner.js');
 let assert = require('chai').assert;
-
+let Recipe = require('../new-recipe.js');
 let systemParticles = require('../system-particles.js');
+let ConvertConstraintsToConnections = require('../strategies/convert-constraints-to-connections.js');
+
 var loader = new Loader();
 systemParticles.register(loader);
 let Person = loader.loadEntity("Person");
@@ -26,7 +28,7 @@ describe('Planner', function() {
     var a = new Arc({id: "test-plan-arc", loader});
     var p = new Planner();
     var population = await(p.plan(a));
-    assert.equal(4, population.length);
+    assert.equal(9, population.length);
   });
 
   it('make a plan with views', async () => {
@@ -36,10 +38,26 @@ describe('Planner', function() {
 
     var p = new Planner();
     var population = await(p.plan(a));
-    assert.equal(7, population.length);
-    console.log(population.length);
-    for (var p of population) {
-      console.log(p.result.toString());
-    }
+    assert.equal(12, population.length);
+  });
+});
+
+describe('ConvertConstraintsToConnections', async() => {
+
+  it('fills out an empty constraint', async() => {
+    var recipe = new Recipe();
+    recipe.newConnectionConstraint('A', 'b', 'C', 'd');
+    var strategizer = {generated: [{result: recipe, score: 1}]};
+    var cctc = new ConvertConstraintsToConnections();
+    let { results } = await cctc.generate(strategizer);
+    assert(results.length == 1);
+    let { result, score } = results[0];
+    assert.deepEqual(result.toString(),
+`recipe
+  map as view0
+  A as particle0
+    b = view0
+  C as particle1
+    d = view0`);
   });
 });
