+++ strategy-explorer/se-explorer.html
@@ -6,6 +6,7 @@
     <style>
       .generation-box {
         border: 1px solid black;
+        width: 1200px;
       }
     </style>
     <div>
